+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Parsley 2.x changelog
 
 ## master
-- Drop support for undocumented option `eventValidate`.
 - type="number" now follows HTML5 spec. In particular, commas are no longer accepted. (#1037)
+- Calls to validate, isValid, whenValidate, whenValid use named arguments
+  (e.g. validate({force: true, group: 'foo'})). Previous API is supported for
+  isValid/validate but is deprecated.
+- Drop support for undocumented option `eventValidate`.
 
 ## 2.2.0-rc3